Add `tflint.issue` function
