+++ opa/engine.go
@@ -34,10 +34,8 @@ func NewEngine(ret *loader.Result) (*Engine, error) {
 	}, nil
 }
 
-// Result is the result of the query.
-// It corresponds to an element of the Set document generated by the rule
-// in Rego, and is a concept corresponding to tflint.Issue.
-type Result struct {
+// Issue is the result of the query.
+type Issue struct {
 	message  string
 	location hcl.Range
 }
@@ -47,8 +45,7 @@ type Result struct {
 // rego.ResultSet is parsed according to the following conventions:
 //
 // - All rules should be under the "tflint" package
-// - Rule should return a Set document
-// - The elements of the set must be objects consisting of "message" and "location".
+// - Rule should return a tflint.issue()
 //
 // Example:
 //
@@ -57,14 +54,11 @@ type Result struct {
 //	deny_test[issue] {
 //	  [condition]
 //
-//	  issue := {
-//	    "message": "test",
-//	    "location": resource.decl_range
-//	  }
+//	  issue := tflint.issue("not allowed", resource.decl_range)
 //	}
 //
 // ```
-func (e *Engine) RunQuery(rule *Rule, runner tflint.Runner) ([]*Result, error) {
+func (e *Engine) RunQuery(rule *Rule, runner tflint.Runner) ([]*Issue, error) {
 	regoOpts := []func(*rego.Rego){
 		// All rules should be under the "tflint" package
 		rego.Query(fmt.Sprintf("data.tflint.%s", rule.RegoName())),
@@ -88,43 +82,25 @@ func (e *Engine) RunQuery(rule *Rule, runner tflint.Runner) ([]*Result, error) {
 		return nil, err
 	}
 
-	var results []*Result
+	var issues []*Issue
 	for _, result := range rs {
 		for _, expr := range result.Expressions {
 			values, ok := expr.Value.([]any)
 			if !ok {
-				return nil, fmt.Errorf("rule document content is not set, got %T", expr.Value)
+				return nil, fmt.Errorf("issue is not set, got %T", expr.Value)
 			}
 
 			for _, value := range values {
-				ret, err := jsonToResult(value, "issue")
+				ret, err := jsonToIssue(value, "issue")
 				if err != nil {
 					return nil, err
 				}
-				results = append(results, ret)
+				issues = append(issues, ret)
 			}
 		}
 	}
 
-	return results, err
-}
-
-func jsonToResult(in any, path string) (*Result, error) {
-	ret, err := jsonToObject(in, path)
-	if err != nil {
-		return nil, err
-	}
-
-	message, err := jsonToString(ret["message"], fmt.Sprintf("%s.message", path))
-	if err != nil {
-		return nil, err
-	}
-	location, err := jsonToRange(ret["location"], fmt.Sprintf("%s.location", path))
-	if err != nil {
-		return nil, err
-	}
-
-	return &Result{message: message, location: location}, nil
+	return issues, err
 }
 
 type PrintHook struct{}
